backend/requirements.txt
django>=6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.7
gunicorn>=22.0
